Pre and post-step intervention along with freezing graph descendants logic
Whenever we perform an intervention in the post-step regime, all of the
descendants in the causal graph of the entity we intervened upon must
freeze their dynamics for the next dynamics step. Furthermore, re-orienting
a vehicle is a joint intervention on both its position and its orientation if
the vehicle was about to move forward had we not intervened.